requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
python-jose[cryptography]
passlib[bcrypt]
imbalanced-learn
matplotlib
seaborn
pytest
pytest-asyncio
httpx
python-multipart